requirements.txt
Flask
Flask-MySQLdb
PyJWT
